sort projects by prio
